==> Topic Modeling/Topic Modeling Scientific Grant Abstracts/code/1. Requesting NIH Grant Abstracts from RePORTER.ipynb <==
# Topic Modeling Scientific Text
*This template and workflow were developed by Margaret Gratian. This set of notebooks can be used to find topics in scientific text.*

____________________________________
## 1. Requesting NIH Grant Abstracts from RePORTER

**Notebook Goals**
- Demonstrate how to request NIH grant abstracts from the NIH RePORTER API.

**Requirements**
- This notebook uses a module called request_reporter to work with the RePORTER API. It is available with these notebooks in a directory called reporter. For this notebook to work, this notebook must be in the same directory that the reporter directory is in. 
- Please see the README for instructions and recommendations on proper installation in order to use this series of packages.

**Inputs**
- No inputs to this notebook.

**Outputs**

The following is a recommended path for saving your data. If you modify it, be sure to modify the inputs and outputs of subsequent notebooks.

- Output Filepath 1: "../data/reporter_results.csv"
    - The result of the RePORTER API request.

## Import Packages

import pandas as pd

# Import the custom package for working with the RePORTER API
from reporter import request_reporter as rr

## Dataset Development

For this example, we will focus on NCI administered R01 grants in FY24.

# Use the function to request_reporter module to get R01 awards for FY 2024
# Note by default of this function, it will return NCI administered awards.
r01_24_results_df = rr.request_nci_awards_by_year_and_activity_codes([2024], ["R01"])
print(r01_24_results_df.shape)

r01_24_results_df.head()

# The abstracts are available in a column called "abstract_text"
# Preview this column
r01_24_results_df[["abstract_text"]].head()

## Save Outputs

r01_24_results_df.to_csv("../data/reporter_results.csv")
